+++ mpupdater/MediaPlayerUpdater.cs
@@ -14,8 +14,13 @@ namespace mpupdater
 {
 	public sealed class MediaPlayerUpdater : Updater
 	{
+#if WIN64
+		const string MediaPlayerPath = @"MPC-HC64\";
+		const string MediaPlayerExecutable = "mpc-hc64.exe";
+#else
 		const string MediaPlayerPath = @"MPC-HC\";
 		const string MediaPlayerExecutable = "mpc-hc.exe";
+#endif
 
 		private const string UpdateURL = "http://nightly.mpc-hc.org/";
 		protected override string GetUpdateURL
@@ -34,6 +39,15 @@ namespace mpupdater
 			}
 		}
 
+#if WIN64
+		protected override string GetVersionRegexPattern
+		{
+			get
+			{
+				return @"MPC-HC\.(\d+)\.(\d+)\.(\d+)(?:\.(\d+))?\.x64.7z";
+			}
+		}
+#else
 		protected override string GetVersionRegexPattern
 		{
 			get
@@ -41,6 +55,7 @@ namespace mpupdater
 				return @"MPC-HC\.(\d+)\.(\d+)\.(\d+)(?:\.(\d+))?\.x86.7z";
 			}
 		}
+#endif
 
 		protected override void GetInstalledVersion()
 		{
@@ -58,7 +73,11 @@ namespace mpupdater
 			if (!CheckUpdate())
 				return;
 
+#if WIN64
+			string fileName = "MPC-HC." + CurrentVersion + ".x64.7z";
+#else
 			string fileName = "MPC-HC." + CurrentVersion + ".x86.7z";
+#endif
 
 			Console.WriteLine("Downloading update...");
 			DownloadUpdateWithProgress(fileName);
@@ -72,7 +91,8 @@ namespace mpupdater
 			}
 			finally
 			{
-				File.Delete(fileName);
+				if (File.Exists(fileName))
+					File.Delete(fileName);
 			}
 			
 			Console.WriteLine("Done.");
